+++ test/dommock/DOMMockBuilder.ts
@@ -1,4 +1,4 @@
-// Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2019-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
 import { Substitute } from '@fluffy-spoon/substitute';
@@ -670,7 +670,7 @@ export default class DOMMockBuilder {
         } as RTCRtpSendParameters;
       }
 
-      replaceTrack(track?: MediaStreamTrack): Promise<void> {
+      async replaceTrack(track?: MediaStreamTrack): Promise<void> {
         return new Promise(resolve => {
           asyncWait(() => {
             this.track = track;