+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow to specify manual version in publish script
 
 ### Fixed
+- Fix DefaultTransceiverController async function signature
+- Make DefaultBrowserBehavior implement BrowserBehavior interface
 - Fix publish script to use npm version
 - Add stage to saucelabs session name for integration tests